custom dices and font size timer
