+++ dices/ITA-7X5.txt
@@ -1,0 +1,36 @@
+7X5
+Qu,O,M,B,A,O
+R,L,E,O,I,U
+C,O,A,I,R,F
+E,P,U,L,T,S
+T,O,I,D,V,M
+C,M,L,D,I,O
+H,N,S,R,E,I
+T,A,C,G,I,P
+E,D,O,N,T,S
+L,I,B,A,R,T
+G,F,A,I,R,P
+O,E,G,U,N,L
+U,C,O,T,N,E
+V,E,T,I,G,N
+E,A,T,I,O,A
+M,A,D,E,P,C
+E,H,I,F,S,E
+B,F,L,O,N,E
+A,L,E,R,C,S
+I,R,O,M,L,C
+C,P,V,A,S,G
+Z,E,V,A,N,D
+O,M,A,R,S,I
+Qu,S,B,Z,F,A
+O,C,A,M,I,B
+U,T,E,S,L,P
+R,A,T,I,B,L
+H,I,S,E,R,N
+M,P,A,C,E,D
+C,O,A,I,F,R
+A,M,S,R,I,O
+V,I,T,E,G,N
+D,A,N,E,Z,V
+N,O,E,U,C,T
+A,M,O,Qu,O,B
